+++ .github/workflows/publish_bedspec.yml
@@ -2,7 +2,7 @@ name: publish
 
 on:
   push:
-    tags: '\d+.\d+.\d+'
+    tags: '[0-9]+.[0-9]+.[0-9]+'
 
 env:
   POETRY_VERSION: 1.6
@@ -16,6 +16,7 @@ jobs:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
+          submodules: true
 
       - uses: rickstaa/action-contains-tag@v1
         id: contains_tag
@@ -23,20 +24,20 @@ jobs:
           reference: "main"
           tag: "${{ github.ref_name }}"
 
-  tests:
-    name: tests
+  unit-tests:
+    name: unit-tests
     needs: on-main-branch-check
     if: ${{ needs.on-main-branch-check.outputs.on_main == 'true' }}
     uses: "./.github/workflows/tests.yml"
 
   build-wheels:
     name: build wheels
-    needs: tests
+    needs: unit-tests
     uses: "./.github/workflows/wheels.yml"
 
   build-sdist:
     name: build source distribution
-    needs: tests
+    needs: unit-tests
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -127,6 +128,6 @@ jobs:
         with:
           name: ${{ github.ref_name }}
           body: |
-            ${{ needs.draft-changelog.outputs.release_body }}
+            ${{ needs.make-changelog.outputs.release_body }}
           draft: false
           prerelease: false
