+++ .github/workflows/tests.yml
@@ -1,4 +1,4 @@
-name: Code checks
+name: unit tests
 
 on:
   push:
@@ -16,7 +16,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        PYTHON_VERSION: ["3.11", "3.12"]
+        PYTHON_VERSION: ["3.10", "3.11", "3.12"]
     steps:
       - uses: actions/checkout@v4
         with:
