+++ bedspec/__init__.py
@@ -1,4 +1,3 @@
-# ruff: noqa: F401
 from ._bedspec import Bed2
 from ._bedspec import Bed3
 from ._bedspec import Bed4
@@ -10,11 +9,35 @@ from ._bedspec import BedGraph
 from ._bedspec import BedLike
 from ._bedspec import BedPE
 from ._bedspec import BedStrand
-from ._bedspec import GenomicSpan
+from ._bedspec import BedType
 from ._bedspec import Named
 from ._bedspec import PairBed
 from ._bedspec import PointBed
+from ._bedspec import ReferenceSpan
 from ._bedspec import SimpleBed
 from ._bedspec import Stranded
-from ._io import BedReader
-from ._io import BedWriter
+from ._reader import BedReader
+from ._writer import BedWriter
+
+__all__ = [
+    "Bed2",
+    "Bed3",
+    "Bed4",
+    "Bed5",
+    "Bed6",
+    "Bed12",
+    "BedColor",
+    "BedGraph",
+    "BedLike",
+    "BedPE",
+    "BedStrand",
+    "BedType",
+    "Named",
+    "PairBed",
+    "PointBed",
+    "ReferenceSpan",
+    "SimpleBed",
+    "Stranded",
+    "BedReader",
+    "BedWriter",
+]
