+++ bedspec/overlap/__init__.py
@@ -1,3 +1,5 @@
-# ruff: noqa: F401
 from ._overlap import OverlapDetector
-from ._overlap import QueryReferenceSpanType
+
+__all__ = [
+    "OverlapDetector",
+]
